+++ sim/bfin/dv-bfin_dma.c
@@ -331,7 +331,7 @@ bfin_dma_io_write_buffer (struct hw *me, const void *source, int space,
     value = dv_load_2 (source);
 
   mmr_off = addr % dma->base;
-  valuep = (void *)((unsigned long)dma + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)dma + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -429,7 +429,7 @@ bfin_dma_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr % dma->base;
-  valuep = (void *)((unsigned long)dma + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)dma + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
