+++ sim/bfin/dv-bfin_ebiu_ddrc.c
@@ -79,7 +79,7 @@ bfin_ebiu_ddrc_io_write_buffer (struct hw *me, const void *source,
     value = dv_load_2 (source);
 
   mmr_off = addr - ddrc->base;
-  valuep = (void *)((unsigned long)ddrc + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)ddrc + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -118,7 +118,7 @@ bfin_ebiu_ddrc_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - ddrc->base;
-  valuep = (void *)((unsigned long)ddrc + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)ddrc + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
