+++ sim/bfin/dv-bfin_gpio.c
@@ -117,7 +117,7 @@ bfin_gpio_io_write_buffer (struct hw *me, const void *source, int space,
 
   value = dv_load_2 (source);
   mmr_off = addr - port->base;
-  valuep = (void *)((unsigned long)port + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)port + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -190,7 +190,7 @@ bfin_gpio_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - port->base;
-  valuep = (void *)((unsigned long)port + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)port + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
