+++ sim/bfin/dv-bfin_otp.c
@@ -116,7 +116,7 @@ bfin_otp_io_write_buffer (struct hw *me, const void *source, int space,
     value = dv_load_2 (source);
 
   mmr_off = addr - otp->base;
-  valuep = (void *)((unsigned long)otp + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)otp + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -192,7 +192,7 @@ bfin_otp_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - otp->base;
-  valuep = (void *)((unsigned long)otp + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)otp + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -280,7 +280,7 @@ bfin_otp_finish (struct hw *me)
   otp->timing  = 0x00001485;
 
   /* Semi-random value for unique chip id.  */
-  bfin_otp_write_page_val2 (otp, FPS00, (unsigned long)otp, ~(unsigned long)otp);
+  bfin_otp_write_page_val2 (otp, FPS00, (uintptr_t)otp, ~(uintptr_t)otp);
 
   memset (part_str, 0, sizeof (part_str));
   sprintf (part_str, "ADSP-BF%iX", type);
