+++ sim/bfin/dv-bfin_pfmon.c
@@ -60,7 +60,7 @@ bfin_pfmon_io_write_buffer (struct hw *me, const void *source, int space,
 
   value = dv_load_4 (source);
   mmr_off = addr - pfmon->base;
-  valuep = (void *)((unsigned long)pfmon + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)pfmon + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -93,7 +93,7 @@ bfin_pfmon_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - pfmon->base;
-  valuep = (void *)((unsigned long)pfmon + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)pfmon + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
