+++ sim/bfin/dv-bfin_pint.c
@@ -74,7 +74,7 @@ bfin_pint_io_write_buffer (struct hw *me, const void *source, int space,
   else
     value = dv_load_2 (source);
   mmr_off = addr - pint->base;
-  valuep = (void *)((unsigned long)pint + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)pint + mmr_base() + mmr_off);
 
   HW_TRACE_WRITE ();
 
@@ -143,7 +143,7 @@ bfin_pint_io_read_buffer (struct hw *me, void *dest, int space,
     return 0;
 
   mmr_off = addr - pint->base;
-  valuep = (void *)((unsigned long)pint + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)pint + mmr_base() + mmr_off);
 
   HW_TRACE_READ ();
