+++ sim/bfin/dv-bfin_pll.c
@@ -70,7 +70,7 @@ bfin_pll_io_write_buffer (struct hw *me, const void *source,
     value = dv_load_2 (source);
 
   mmr_off = addr - pll->base;
-  valuep = (void *)((unsigned long)pll + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)pll + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -109,7 +109,7 @@ bfin_pll_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - pll->base;
-  valuep = (void *)((unsigned long)pll + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)pll + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
