+++ sim/bfin/dv-bfin_rtc.c
@@ -72,7 +72,7 @@ bfin_rtc_io_write_buffer (struct hw *me, const void *source,
     value = dv_load_2 (source);
 
   mmr_off = addr - rtc->base;
-  valuep = (void *)((unsigned long)rtc + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)rtc + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
 
@@ -114,7 +114,7 @@ bfin_rtc_io_read_buffer (struct hw *me, void *dest,
     return 0;
 
   mmr_off = addr - rtc->base;
-  valuep = (void *)((unsigned long)rtc + mmr_base() + mmr_off);
+  valuep = (void *)((uintptr_t)rtc + mmr_base() + mmr_off);
   value16p = valuep;
   value32p = valuep;
